Broke out SubmissionModActions.java from PopulateSubmissionViewHolder.java
