+++ app/src/main/res/values/strings.xml
@@ -1312,4 +1312,16 @@
     <string name="settings_dialog_colored_border_desc">Display a thin colored border around some dialog boxes.</string>
 
     <string name="misc_err_sync_failed">Failed to sync subreddits. No subscribed subreddits?</string>
+
+    <string name="mod_btn_distinguish">Distinguish</string>
+    <string name="mod_btn_undistinguish">Undistingiush</string>
+
+    <string name="nsfw_status_removed">NSFW status removed</string>
+    <string name="nsfw_status_set">NSFW status set</string>
+
+    <string name="spoiler_status_removed">Spoiler status removed</string>
+    <string name="spoiler_status_set">Spoiler status set</string>
+
+    <string name="submission_distinguished">Submission distinguished</string>
+    <string name="submission_distinguished_removed">Submission distinguish removed</string>
 </resources>
